+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Android push notification taps open the app.
+- Push notification taps refresh friends.
 
 ## [1.2.1] - 2026-04-13
 ### Fixed